refactor: resolve dependencies
