+++ pkg-manager/resolve-dependencies/test/hoistPeers.test.ts
@@ -1,4 +1,4 @@
-import { hoistPeers } from '../lib/hoistPeers'
+import { hoistPeers, hoistOptionalPeers } from '../lib/hoistPeers'
 
 test('hoistPeers picks an already available prerelease version', () => {
   expect(hoistPeers([['foo', { range: '*' }]], {
@@ -10,5 +10,20 @@ test('hoistPeers picks an already available prerelease version', () => {
     },
   })).toStrictEqual({
     foo: '1.0.0-beta.0',
+  })
+})
+
+test('hoistOptionalPeers only picks a version that satisfies all optional ranges', () => {
+  expect(hoistOptionalPeers({
+    foo: ['2', '2.1'],
+  }, {
+    foo: {
+      '1.0.0': 'version',
+      '2.0.0': 'version',
+      '2.1.0': 'version',
+      '3.0.0': 'version',
+    },
+  })).toStrictEqual({
+    foo: '2.1.0',
   })
 })
